feat(apple): 共用云端目录与候选调频管理界面
