+++ cmake/BackendAccount.cmake
@@ -12,6 +12,10 @@ set(MSIME_ACCOUNT_SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/shared/backend/BackendPreferencesClient.swift"
     "${CMAKE_CURRENT_SOURCE_DIR}/shared/backend/BackendDictionaryClient.swift"
     "${CMAKE_CURRENT_SOURCE_DIR}/shared/backend/BackendCandidateClient.swift"
+    "${CMAKE_CURRENT_SOURCE_DIR}/shared/backend-ui/BackendInputModifiers.swift"
+    "${CMAKE_CURRENT_SOURCE_DIR}/shared/backend-ui/CloudCandidatesView.swift"
+    "${CMAKE_CURRENT_SOURCE_DIR}/shared/backend-ui/CloudDictionaryCatalogView.swift"
+    "${CMAKE_CURRENT_SOURCE_DIR}/shared/backend-ui/CloudDictionaryEditor.swift"
     "${METASEQUOIA_MACOS_ROOT}/src/BackendDictionaryView.swift"
     "${METASEQUOIA_MACOS_ROOT}/src/BackendSettingsView.swift"
     "${METASEQUOIA_MACOS_ROOT}/src/BackendClipboardView.swift"
