+++ platforms/ios/App/Sources/CloudDictionaryView.swift
@@ -134,56 +134,3 @@ struct CloudDictionaryView: View {
     }
   }
 }
-
-struct CloudDictionaryEditor: View {
-  let kind: BackendAccountClient.DictionaryKind
-  let isEditing: Bool
-  let save: (BackendAccountClient.DictionaryValue) async throws -> Void
-  @Environment(\.dismiss) private var dismiss
-  @State private var code: String
-  @State private var word: String
-  @State private var weight: String
-  @State private var message: String?
-  @State private var busy = false
-  @State private var pending: Task<Void, Never>?
-  init(kind: BackendAccountClient.DictionaryKind, value: BackendAccountClient.DictionaryValue?, save: @escaping (BackendAccountClient.DictionaryValue) async throws -> Void) {
-    self.kind = kind; self.isEditing = value != nil; self.save = save
-    _code = State(initialValue: value?.code ?? "")
-    _word = State(initialValue: value?.word ?? "")
-    _weight = State(initialValue: String(value?.weight ?? 100_000))
-  }
-  var body: some View {
-    NavigationView {
-      Form {
-        Section(kind.title) {
-          TextField("词条内容", text: $word)
-          TextField("完整编码", text: $code).textInputAutocapitalization(.never).autocorrectionDisabled()
-          TextField("权重", text: $weight).keyboardType(.numberPad)
-        }
-        Text("点击上传将把此词条保存到当前账号。版本冲突或重复词条时需返回刷新，再重新确认。")
-          .font(.footnote).foregroundStyle(.secondary)
-        if let message { Text(message).foregroundStyle(.red) }
-        if busy { ProgressView() }
-      }
-      .navigationTitle(isEditing ? "编辑云端词条" : "添加云端词条")
-      .disabled(busy)
-      .toolbar {
-        ToolbarItem(placement: .cancellationAction) { Button("取消") { pending?.cancel(); dismiss() } }
-        ToolbarItem(placement: .confirmationAction) {
-          Button("上传") {
-            guard let number = Int64(weight), number >= 0, !word.isEmpty else { message = "请填写词条和有效的非负权重。"; return }
-            busy = true; message = nil
-            pending = Task {
-              defer { busy = false }
-              do { try await save(.init(code: code, word: word, weight: number)); try Task.checkCancellation(); dismiss() }
-              catch is CancellationError { }
-              catch { message = error.localizedDescription }
-            }
-          }.disabled(busy)
-        }
-      }
-    }
-    .interactiveDismissDisabled(busy)
-    .onDisappear { pending?.cancel() }
-  }
-}
