+++ platforms/ios/project.yml
@@ -91,6 +91,7 @@ targets:
     platform: iOS
     sources:
       - path: platforms/ios/App/Sources
+      - path: shared/backend-ui
       - path: shared/backend/BackendAccountClient.swift
       - path: shared/backend/BackendChatClient.swift
       - path: shared/backend/BackendAccountSession.swift
